api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pymysql==1.1.1
DBUtils==3.1.0
python-dotenv==1.0.1
groq>=1.1.2
pdfplumber==0.11.0
python-docx==1.1.2
Pillow==10.3.0
pytesseract==0.3.10
sentence-transformers>=2.7.0
scikit-learn>=1.4.0
numpy>=1.24.0
